+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/nationalarchives/tna-frontend/compare/v0.31.0...HEAD)
 
 ### Added
+
+- Scrollable tables can now have sticky first columns using `tna-table-wrapper--sticky-first-column`
+
 ### Changed
 ### Deprecated
 ### Removed